Rename blackbox test files under monkey_island/cc/resources/blackbox/
