+++ monkey/monkey_island/cc/resources/blackbox/log_blackbox_endpoint.py
