+++ monkey/monkey_island/cc/resources/blackbox/monkey_blackbox_endpoint.py
