+++ monkey/monkey_island/cc/resources/blackbox/telemetry_blackbox_endpoint.py
